Simplified DB access.
Previously we used to check if options was passed into our DBContext to decided what database to use.
Now we let our appsetting.json configuration decide, I've personally setup an userSecret.json, that will be default on MY enviroment.

Also commented out all coupled code in the project, so that it will run.